Refactor out repojection to different file
