+++ ocf_datapipes/load/topographic.py
@@ -1,12 +1,7 @@
 from pathlib import Path
 from typing import Union
 
-import cartopy.crs as ccrs
-import numpy as np
-import pyproj
-import pyresample
 import rioxarray
-import xarray as xr
 from torchdata.datapipes import functional_datapipe
 from torchdata.datapipes.iter import IterDataPipe
 
@@ -28,92 +23,5 @@ class OpenTopographyIterDataPipe(IterDataPipe):
         # Use our standard naming:
         topo = topo.rename({"x": "x_osgb", "y": "y_osgb"})
 
-        # Select Western Europe:
-        topo = topo.sel(x_osgb=slice(-300_000, 1_500_000), y_osgb=slice(1_300_000, -800_000))
-
-        topo = reproject_topo_data_from_osgb_to_geostationary(topo)
-        topo = topo.fillna(0)
-
         while True:
             yield topo
-
-
-def reproject_topo_data_from_osgb_to_geostationary(topo: xr.DataArray) -> xr.DataArray:
-    topo_osgb_area_def = _get_topo_osgb_area_def(topo)
-    topo_geostationary_area_def = _get_topo_geostationary_area_def(topo)
-    topo_image = pyresample.image.ImageContainerQuick(topo.values, topo_osgb_area_def)
-    topo_image_resampled = topo_image.resample(topo_geostationary_area_def)
-    return _get_data_array_of_resampled_topo_image(topo_image_resampled)
-
-
-def _get_topo_osgb_area_def(topo: xr.DataArray) -> pyresample.geometry.AreaDefinition:
-    # Get AreaDefinition of the OSGB topographical data:
-    osgb = ccrs.OSGB(approx=False)
-    return pyresample.create_area_def(
-        area_id="OSGB",
-        projection=osgb.proj4_params,
-        shape=topo.shape,  # y, x
-        area_extent=(
-            topo.x_osgb[0].item(),  # lower_left_x
-            topo.y_osgb[-1].item(),  # lower_left_y
-            topo.x_osgb[-1].item(),  # upper_right_x
-            topo.y_osgb[0].item(),  # upper_right_y
-        ),
-    )
-
-
-def _get_topo_geostationary_area_def(topo: xr.DataArray) -> pyresample.geometry.AreaDefinition:
-    # Get the geostationary boundaries of the topo data:
-    OSGB_EPSG_CODE = 27700
-    GEOSTATIONARY_PROJ = {
-        "proj": "geos",
-        "lon_0": 9.5,
-        "h": 35785831,
-        "x_0": 0,
-        "y_0": 0,
-        "a": 6378169,
-        "rf": 295.488065897014,
-        "no_defs": None,
-        "type": "crs",
-    }
-    osgb_to_geostationary = pyproj.Transformer.from_crs(
-        crs_from=OSGB_EPSG_CODE, crs_to=GEOSTATIONARY_PROJ
-    ).transform
-    lower_left_geos = osgb_to_geostationary(xx=topo.x_osgb[0], yy=topo.y_osgb[-1])
-    upper_right_geos = osgb_to_geostationary(xx=topo.x_osgb[-1], yy=topo.y_osgb[0])
-    shape = (topo.shape[0] * 2, topo.shape[1] * 2)  # Oversample to make sure we don't loose info.
-    return pyresample.create_area_def(
-        area_id="msg_seviri_rss_1km",
-        projection=GEOSTATIONARY_PROJ,
-        shape=shape,
-        # lower_left_x, lower_left_y, upper_right_x, upper_right_y:
-        area_extent=lower_left_geos + upper_right_geos,
-    )
-
-
-def _get_data_array_of_resampled_topo_image(
-    topo_image_resampled: pyresample.image.ImageContainer,
-) -> xr.DataArray:
-    (
-        lower_left_x,
-        lower_left_y,
-        upper_right_x,
-        upper_right_y,
-    ) = topo_image_resampled.geo_def.area_extent
-    return xr.DataArray(
-        topo_image_resampled.image_data,
-        coords=(
-            (
-                "y",
-                np.linspace(
-                    start=upper_right_y, stop=lower_left_y, num=topo_image_resampled.shape[0]
-                ),
-            ),
-            (
-                "x",
-                np.linspace(
-                    start=lower_left_x, stop=upper_right_x, num=topo_image_resampled.shape[1]
-                ),
-            ),
-        ),
-    )
