+++ ocf_datapipes/transform/xarray/__init__.py
@@ -9,3 +9,4 @@ from .downsample import DownsampleIterDataPipe as Downsample
 from .normalize import NormalizeIterDataPipe as Normalize
 from .pv_power_rolling_window import PVPowerRollingWindowIterDataPipe as PVPowerRollingWindow
 from .reduce_num_pv_systems import ReduceNumPVSystemsIterDataPipe as ReduceNumPVSystems
+from .reproject_topographic_data import ReprojectTopographyIterDataPipe as ReprojectTopography
