+++ requirements.txt
@@ -18,3 +18,7 @@ jpeg_xl_float_with_nans
 h5netcdf
 tqdm
 bottleneck
+cartopy
+pyproj
+pyresample
+
